Clean-up write_film_grain_params()
Various refactoring and clean-up to write_film_grain_params():
* Instead of the confusing logic with flip_back_update_parameters_flag,
  simply copy current_frame->film_grain_params to
  cur_frame->film_grain_params in finalize_encoded_frame before setting
  the update_parameters flag appropriately.
* Iterate the film grain parameters random seed in
  finalize_encoded_frame instead of write_film_grain_params
* Previously a few assignments silently fixed erroneous/meaningless
  state/configuration: replace these with assertions to make it clearer
  what's happening.

After these changes write_film_grain_params() can take a const AV1_COMP
and only has the effect of writing to the bitstream.  This is part of my
work at removing side effects and other unexpected actions from
av1_pack_bitstream and below, which is itself part of wider
restructuring and refactoring with the goal of achieving a clean API
separation at the entry to the low-level encoder.

BUG=aomedia:2244

Change-Id: I0162b0d48c3207212d6ce7ce6d7a6d8015294241