Consolidate simple_motion_search based models
This patch creates a model based on features from simple_motion_search
and outputs the class probability for the 10 partition types.

Currently this replaces the model in simple_motion_search_prune_rect
with a relaxed threshold. But a better performance should be achievable
by fine-tuning the thresholds.

Performance:
 OVR_PSNR | SPEED_UP | SPEED_UP:OVR_PSNR |
  0.038%  |  1.783%  |      46.921:1     |

The PSNR is averaged over 60 frames on midres. The speed up is averaged
over30 frames on 4 midres clips with 5 bitrates each.

STATS_CHANGED

Change-Id: I4d64dab6713867d8f9af06d10420073bdf25c5a4