Populate rd_stats appropriately
Populate rd_stats appropriately in the case of explicit skip.

For speed = 1 and 4 presets, BD-rate impact is seen as -0.05% and 0.01% (as per AWCY runs),
For speeds 2 and 3 no BD-rate impact is seen.
No speed impact is seen for presets 1 and 2, for presets = 3 and 4 encode time reduction
is seen as 0.2% and 0.3% (averaged across multiple test cases) respectively.

STATS_CHANGED

Change-Id: I14087831d04f03f0eea7482914809e42db957489