Update patch set 2
Patch Set 2: Code-Review+2

This is an important fix that we should get in as soon as possible.

I checked this CL carefully and verified that it reverts the changes from the
public headers in the aom/aom/ directory. The only exception is the definition
of the following macros in aom/aom/aom_integer.h, which we can deal with in a
separate CL:

+/*!\brief force enum to be unsigned 1 byte*/
+#define UENUM1BYTE(enumvar) \
+  ;                         \
+  typedef uint8_t enumvar
+
+/*!\brief force enum to be signed 1 byte*/
+#define SENUM1BYTE(enumvar) \
+  ;                         \
+  typedef int8_t enumvar
+
+/*!\brief force enum to be unsigned 2 byte*/
+#define UENUM2BYTE(enumvar) \
+  ;                         \
+  typedef uint16_t enumvar
+
+/*!\brief force enum to be signed 2 byte*/
+#define SENUM2BYTE(enumvar) \
+  ;                         \
+  typedef int16_t enumvar
+
+/*!\brief force enum to be unsigned 4 byte*/
+#define UENUM4BYTE(enumvar) \
+  ;                         \
+  typedef uint32_t enumvar
+
+/*!\brief force enum to be unsigned 4 byte*/
+#define SENUM4BYTE(enumvar) \
+  ;                         \
+  typedef int32_t enumvar
+

Patch-set: 2
Reviewer: Gerrit User 9545 <9545@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+2