Update patch set 12
Patch Set 12:

(1 comment)

Currently our optimization is applicable for Dual filter(speed = 1).
we are still working for speed > 1 for non-dual filters.

In case of screen contents we expect regular filter to win most of the times. With our implementations regular filter may be forced more no. of times as compared to parent version, hence entropy modelling is possibly favored in our implementations.

In case of VP9 this speed feature has ~3.1% encoder time impact for speed=3, however this is similar to non-dual implementation in av1, hence we are seeing lesser gains in av1 for speed=1(as both horizontal and vertical filters of left and top block should match).
For speed >=2 we expect similar to that of VP9 but the BD-Rate impact needs to be assessed.

Patch-set: 12
Reviewer: Gerrit User 9670 <9670@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>