Merge encode_{with,without}_recode_loop()
There is a lot of overlap between encode_with_recode_loop() and
encode_without_recode_loop(): Since the two were split, a lot of
duplicate code has been added to both functions.  In this commit I merge
the two functions so that encode_with_recode_loop is always used and
recoding can be disabled by setting sf.recode_loop = DISALLOW_RECODE.  I
feel this simplifies the code and it should not have any significant
speed implication: the extra complexity in encode_with_recode_loop() is
suppressed if recoding is disallowed.

At the same time, I move all the code concerned with updating q in the
recode loop to a new function, recode_loop_update_q(), so that
encode_with_recode_loop() is less vast.

I also remove a commented-out call to update_base_skip_probs(), which
has not existed since 2013.

This forms part of wider restructuring and refactoring in order to
achieve a clean API separation at the entry to the low-level encoder.

BUG=aomedia:2244

Change-Id: Ib41f593325c82015ba87e64226a923ffba73cf68