Update patch set 2
Patch Set 2:

Note that the specification only defines the decoding process for valid bitstreams when starting from the start of a coded video sequence.

If the libaom decoder is used on an invalid bitstream or starts decoding not at a keyframe, then the assert might fire.  It may be better to keep the original code here as this would allow the decoder to proceed (of course the decoded frame is quite likely to be incorrect, but later frames may become correct).

Patch-set: 2