Specialize the curvfit / surface fit rate models
The rate models for the curvfit and surffit models are
specialized by block size, by dividing the block sizes
into 4 categories. The prediction error for rate is
reduced substantially in the process.

Note there is much less improvement in distortion by
the same process and so a generic model for all block
sizes are still used for distortion. However the
distortion model for surffit is also updated by sampling
a little more coarsely than before.

STATS_CHANGED

Change-Id: I1ef7335b2e40308e11790dd9fc13d062d4fc81c1