Assert the precondition for update_ref_frame_id().
Assert frame_id_numbers_present_flag is true in update_ref_frame_id().
update_ref_frame_id() uses the valid_for_referencing array, equivalent
to the RefValid array in the spec. This assertion helps ensure that
valid_for_referencing is used only when frame_id_numbers_present_flag is
true, a property described in the Note in Section 6.8.3 (Reference frame
marking semantics) of the spec:

  Note: RefValid is only used to define valid bitstreams when
  frame_id_numbers_present_flag is equal to 1.  ...

This property is satisfied by the libaom reference code, but is not
fully satisfied by the spec.

BUG=2295

Change-Id: I4438d86b1b645304b627bc1f7d241f1831435253