Update patch set 3
Patch Set 3:

Copied from Ranjit's email:
We expect it to help speed=5 as well. We did not disable the logic for speed >=5 because of error we saw for patchset1 (for speed =8).

For reference, here is the error for patchset1:
/var/jenkins/workspace/libaom__per_commit-compile_sanitizers/sanitizer/address/aom/test/horz_superres_test.cc:164
Expected: (psnr) > (GetPsnrThreshold()), actual: 25.929 vs 26
superres_mode_ = , superres_denom_ = 9, superres_kf_denom_ = 9

Before patchset 1, psnr is 26.283, and after it, psnr is 25.929 that is lower than the threshold.

The above test is only for speed 8, that is not a well-tuned speed.

It might be better to figure out why this feature causes big BD rate changes(worse for speed 4, but better for speed >4) instead of having this work around. What do you think?

Patch-set: 3
Reviewer: Gerrit User 5185 <5185@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>