Move per-frame speed-level into EncodeFrameParams
Move speed-level into EncodeFrameParams so that, in theory, a different
speed-level can be set per frame (though this is currently untested and
the per-frame speed level is just set to oxcf->speed).

This forms part of wider restructuring and refactoring in order to
achieve a clean API separation at the entry to the low-level encoder.

BUG=aomedia:2244

Change-Id: I54f81ecb8ffa01bac26b406305108553b17ceea7