Move check_show_existing_frame to next frame
Before this patch, check_show_existing_frame runs at the end of encoding
a frame to check if the *next* frame should be a show_existing_frame.
This patch changes the code so that check_show_existing_frame() runs in
the strategy *before* encoding a frame to check if *this* frame should
be a show_existing_frame.  As well as being less confusing this will
make it a lot easier for show_existing_frame to be moved to
EncodeFrameParams.

Also, remove effectless assignments to rc.is_bwd_ref_frame and
rc.is_last_bipred_frame

This forms part of wider restructuring and refactoring in order to
achieve a clean API separation at the entry to the low-level encoder and
to make the high-level encoding strategy easier to modify.

BUG=aomedia:2244

Change-Id: I9fab1c547bf681c321901f7e3548d811ee8306f8