Update patch set 9
Patch Set 9:

All, 
We tried repeating the AWCY run for a given recipe and version for (speed=4) twice and the same can be found at the link below:
https://beta.arewecompressedyet.com/?job=Jan25_ref_c4%402019-01-25T09%3A03%3A23.526Z&job=Jan25_ref_c4_1%402019-01-28T17%3A47%3A30.883Z

We found that at higher bit-rates, encode time variation is as high as ~2% (i.e., run to run variation). We did not find much variation in encode time at low bit-rates, though (as per the current run).

The speed data we reported in commit message is obtained as an average of encode time reduction obtained for  - 30 frames of BasketBallDrill_832x480_50 at 1 mbps,
30 frames of crowd_run_640x360_10 at 750 kbps and 15 frames of parkrun_720p50 at qp=50.

Regards,
Ranjit

Patch-set: 9