Update patch set 9
Patch Set 9:

> Hi Tom, do you know how AWCY gets encoder/decoder timing data and are the timing data reliable? Thanks.

Well, i don't know the inner workings of AWCY, and as Ranjit said, there are probably 1-2% inaccuracy in the measurements, but >5% improvement is pretty significant, so AWCY should show a clear improvement in the encoding times. It doesn't do that in this case. Also, AWCY averages many different videos at different bitrates, so it should indicate the overall improvement more accurately than a few individual tests in my opinion.

Patch-set: 9