setup_buffer_ref_mvs_inter: bugfix for scaled refs
A recent patch (https://aomedia-review.googlesource.com/c/aom/+/78181)
fixed a crash for superres + resize case reported in the bug referenced
below.

However, that was a partial fix, and it resulted in a very large BDRate
(PSNR) drop for superres (~50% drop) use case.

This patch adds the missing part to complete the patch:
- We use the scaled reference only for the purpose of av1_mv_pred()
- Then use the unscaled reference to re-setup 'yv12_mb' for subsequent
use.

As a result:
- We fully recover the BDRate loss for superres case (tested both
SUPERRES_FIXED and SUPERRES_QTHRESH modes).
- The superres + resize case from the bug also tested OK (no crash).

BUG=aomedia:2293

Change-Id: I37990c54c6d1842b217e133981c7039f1c37df90