Optimize multipliers for rd-thresh
STATS_CHANGED: changed the thresholds for when a mode should be skipped;
made small changes to each constant and then observed the drop/increase in
PSNR and encode time. Combine, these changes have a PSNR drop of 0.007 on the
lowrest test set, and a 1% speedup on first 30 frames of
BasketballDrill_832x480_50.y4m.

Change-Id: I4bf5c945ea48094f4d678a1cbfc74cf13db4c85e