Add speed feature to disable loop restoration search for chroma
This patch will introduce a speed feature to disable the loop
restoration search for chroma plane for all frames except boosted
frames and when screen content tool is on. This speed feature is
enabled for preset 3 and above.

For speed = 3 and 4 presets, BD-rate drop is seen by 0.01% and
0.01% (for AWCY runs) with encode time reduction of 1.75% and 2.23%
(averaged across multiple test cases) respectively.

STATS_CHANGED

Change-Id: I2d2dab3a0114179f0fdd49c8d44d69a3b372e9f2