Streamline compound types
This patch includes various refactoring towards integrating
dist_wtd compound type search with the other compound modes.

The current patch keeps separate rd loops for distwtd compound
and other compound modes through the macro SEPARATE_COMP_DISTWTD_RD.
Currently it is set as 1 so that no change is introduced in
the coding stats.

Experiments are in progress for turning this macro off and
cleaning up the code.

Change-Id: Ibc29fd4d088f977053cdd38fc03ea2005b073c07