first pass: Respect the rc->max_gf_interval.
max_gf_interval may be provided by the user, or we may have set it to a
default value in av1_rc_init().

So, we should respect that value in the breakout condition, instead of
recalculating it based on pyramid levels.

No change in stats if there was no user provided value for
max_gf_interval.

Change-Id: Ie462ba8b53229dd96d8b7c10a867caf01476f653