Move refresh_*_frame flags to EncodeFrameParams
refresh_*_frame flags are used to tell the low-level encode which
reference frame buffers should be refreshed by this frame.  This patch
moves these flags to EncodeFrameParams to make it clear that they are
set by the high-level strategy code and passed to the low-level encoder
as parameters.  The flags are copied to their original locations by
av1_encode().

This forms part of wider restructuring and refactoring in order to
achieve a clean API separation at the entry to the low-level encoder and
to make the high-level encoding strategy easier to modify.

BUG=aomedia:2244

Change-Id: I71825ab43a0510c5b87a083749b29dbc7a41cc4e