Move refresh_*_frame flags to EncodeFrameParams
refresh_*_frame flags are used to tell the low-level encoder which
reference frame buffers should be refreshed by this frame.  This patch
moves these flags to EncodeFrameParams to make it clear that they are
set by the high-level strategy code and passed to the low-level encoder
as parameters.  The flags are copied to their original locations by
av1_encode().

The primary uses of these in the low-level encoder are:
* By get_refresh_frame_flags() to calculate refresh_frame_flags
* By update_reference_frames() to determine reference buffer assignments
I plan to move these two to high-level strategy imminently so that
refresh_frame_flags and reference buffer assignments are parameters
passed to the low-level encoder.  A variety of other things also have
behaviour depending on these flags, which should probably be changed
to be explicit parameters in EncodeFrameParams.

This forms part of wider restructuring and refactoring in order to
achieve a clean API separation at the entry to the low-level encoder and
to make the high-level encoding strategy easier to modify.

BUG=aomedia:2244

Change-Id: I71825ab43a0510c5b87a083749b29dbc7a41cc4e