Refactor compound search loop
Refactors the two-loop compound search framework to create a
generic framework where any set of modes could be used in the
first pass and the rest in the second pass.

Also whether to use 2-pass or 1-pass is now a speed-feature.
When turned on, it gives 4% speed up with ~0.08% loss, as
tested on lowres, midres and AWCY-Objective-1-Fast.

Change-Id: Iaca3cc6ac36193cbfe34805795c51024df538ee8