Refine interpolation_filter_search
This patch will restructure the match found logic in interpolation_filter_search().
In interpolation stats, rd and sse of interpolation search result is included.
Same values are restored, if current mode is matching with interpolation stats.
In this way, calculation of model_rd is avoided. It also reduces the number of calls
to build_inter_predictor() function for early exit case.

This change is bit exact. For speed = 1, 2, 3 and 4 presets, encode time reduction
is seen as 0.40%, 0.85%, 1.40% and 1.20% (averaged across multiple test cases).

Change-Id: Ie8ace3f758c9da9143cc602ce14b876a7b78f299