Refine interpolation_filter_search
This patch will restructure the match found logic in interpolation_filter_search().
In interpolation stats, rd and sse of interpolation search result is included.
Same values are restored, if current mode is matching with interpolation stats.
In this way, calculation of model_rd is avoided. It also reduces the number of calls
to build_inter_predictor() function for early exit case.

Change-Id: Ie8ace3f758c9da9143cc602ce14b876a7b78f299