Add an optimization for faster compound search
In the case when compound_type_rd() is called with both
average and dist-wtd compound enabled, use the rd model to
first differentiate between the two and then use the
estimate_yrd_for_sb() function only for the better of the
two. This gives about 2% speed-up in single loop
compound search mode.

Change-Id: I05e028ca867a1cc05337cc570aadbb36996283af