Do NOT reduce length of last GF group for q > 128.
Before:
Reducing the GF length was creating pulsing because:
- The last inter frame (before a key frame) used a very low quality
(similar quality as leaf level frames), and
- The next keyframe used a much better quality

This was due to the fact that the last inter frame was NOT an ALT-REF
frame, but in a GF group by itself (along with an overlay frame).

After this patch:
The last inter frame is an ALT-REF frame, so it is at a quality
closer to the subsequent keyframe.
This has a BDRate loss of 0.2%, but the pulsing is resolved.

Note: This change applies to q > 128 only, to get a good compromise
between:
- Reducing pulsing at practical quality settings, and
- Keeping the BDRate loss minimal.

BUG=aomedia:2308

Change-Id: I82ea59e79fa55d2f58f6ddb710d05a88277280ee