Change apostrophe in comments to ASCII character
A lot of warnings caused by this character in comments, while
build by Visual studio if the system language set to a language
other than English (Such as Chinese).

"warning C4819: The file contains a character that cannot
be represented in the current code page (936). Save the file
in Unicode format to prevent data loss"

Change-Id: I9676569c457af8368c432480480a2adf49237ca5