More strict rule to decide if intrabc should be on
In addition to the number of colors in a block, consider also the
variance of the block to decide if intrabc should be enabled. This can
prevent mistakenly predicting some smooth areas as screen content,
where intrabc should actually be disabled to allow loop filters.

Tested keyframe encoding on speed 1:
lowres            -0.04%(-1.44% on flower)
midres            -0.24%(-4.75 on modisode2; -2.45% on snow_mnt)
screen_content    -0.14%(-2.18% on liquify)

BUG=aomedia:2319

Change-Id: If1d0417785a32a6e8ad85757d690645285c802b6