Move interp_filter_selected to RefCntBuffer
cpi->interp_filter_selected is indexed by reference, so a swathe of
memcpys are required to keep it in sync with reference frame refreshing
and reference reassignment.  In this commit I move it to RefCntBuffer so
it is automatically refreshed and and assigned.

Because the old memcpys did not precisely follow reference buffer
refreshing and re-assignment, behaviour is changed slightly by this
patch when the interp filter mask speed-feature is enabled (speed >= 2)
STATS_CHANGED

Change-Id: Iad492bc4b76ded5b57576a19b20c348ab5906415