Refactor reference frame mask.
Earlier, the mask had a few drawbacks:

(i) It only supported disabling a particular reference
entirely (when it is not available, for example).
But it did NOT support allowing/disallowing specific single/compound
reference combinations.

(ii) The mask also had a weird implementation/meaning: to disable a reference
frame, the bit was turned on in mask1, and *all* bits were turned on for
mask2.

Instead, we now use a 2D array that specifies if reference combination
(i,j) is allowed or not. This overcomes both drawbacks above.

No change in stats/output.

Change-Id: Icfa40f24cd6a48775ec0ddfa340019607d0c941c