Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

It seems the ability to comment on the actual patch is not working in my browser any longer, but please allow me to add the comment here:

On line 66, it may not be obvious to someone reading the simplified condition that in the else branch prev_ul >= 0 => prev_u >= 0 && prev_l >= 0. I would recommend adding an assert and/or a comment to that effect.

Patch-set: 1
Reviewer: Gerrit User 5075 <5075@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>