Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+2
> 
> Nice work! This result seems really good!
> 
> Just a quick question, do you have the performance on midres? I would assume the speed up to be a bit less on higher resolution.

On midres, the speedup is 9%, the BDRate is 0.19%, so the ratio is not as good as what it achieves on lowres (midres ratio is around 47x, hdres ratio is around 14x with 0.46% perf loss, even if training the NN separately for different resolution, the ACC decreases by 0.2% with each resolution increase).
In this patch, the feature is made size-dependent, not turned on if encoding midres clips, because I think the ratio is borderline acceptable and 0.19% perf change is too big for speed 0. 
I am playing with thresholds to create a relaxed max size pred mode sf.auto_min_max_partition_based_on_simple_motion = 2.

Patch-set: 3
Reviewer: Gerrit User 5188 <5188@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>