Update patch set 2
Patch Set 2: Code-Review+1

So, if we have noise_level greater than zero, we always set the film_grain_parameters_present to 1 and then decide for each frame whether the parameters are present? This is better than missing the film grain for the whole GOP because it was not found in the first frame.

Patch-set: 2
Reviewer: Gerrit User 6720 <6720@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+1