Add speed feature to disable trellis in estimate_yrd_for_sb
With the default version of quantize_b, this gives 0.09% drop in
BDRate for 7% speed up.

With the adaptive version of quantize_b, this gives 0.03% drop in
BDRate for 7% speed up. Adaptive quant_b still needs SIMD code, so it
cannot yet be used by default for this speed feature.

Not ready to submit, rerunning the testsets afer a refactor and rebase.

Change-Id: I1d64a7545d23243790eba23af724f53bfc495f7d