Fix VisualStudio build
VS C++ compiler doesn't support named initializers. There are multiple
ways to fix the problem, but for now - move the static const array to a
C file where it is actually used. If we will need access for this data
in other parts of the code we can solve it either by providing access
functions or by making an array public in C code and providing C++
access via 'extern'

Change-Id: Ifd28a4db6e9455a45eceea495b47d3d70e26d86f