Bugfix for gf-max-pyr-height option.
oxcf->gf_max_pyr_height was being set to the appropriate value *after*
av1_rc_init() was already called. So, rc->max_gf_length was always 16
earlier, irrespective of gf_max_pyr_height value passed in by the user.

This inconsistency was causing the creation of an invalid bitstream.

So, we now update min and max intervals after each update to
oxcf->gf_max_pyr_height.

Also, for max height = 1 in particular, ensure that bipred is disabled.

With this, max height of 1, 2 and 3 work correctly. (4 was working fine
already)

BUG=aomedia:2345

Change-Id: I0692a99533b2891bc9d6da91c38367a375cb0f3b