Prevent encoder to use wrong distortion
This commit adds a check to compare the distortion computed from pixel
domain with that computed from transform domain for blocks with high
residue energy. For high energy blocks, where there is a tendency of
reconstructed pixel values may have been clamped, encoder uses the
higher one of the two distions computed, this prevents encoder to
choose a mode/transform where resdue engery is super high but pixel
domain distoration is low as a result of clamping in reconstruction.

BUG=aomedia:1973

Change-Id: I9d733b42aa137f0e43559af07f9f648371d4ee96