Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > (1 comment)
> > 
> > Please check, I think I found a bug in "sse_diff = block_sse - this_rd_stats.sse;", which this patch has based on.
> 
> One correction, please ignore the part:
> "To verify this, simply putting "assert(sse_diff >= 0);" right after this line and it will hit.", since sse_diff can be negative anyway (it happens if the distortion is larger than the source variance anyway).

Thanks for the comment, I think the code does reflect my intention. 

block_sse is a sum of squared values of src_diff. Looking from the perspective of the residue block, it is the L2 energy measurement of residue block. 

this_rd_stats.sse is a sum of squared values of transform coefficients. Looking at from the perspective of the residue block, it is also the L2 energy measurement, but in transform domain. However, for 64x64 transform, this only summarized the coefficients in top-left quadrant. 


sse_diff is what I use to measure the of the energy difference between pixel domain(block_sse) and the one transform domain (top left quadrant of 64x64 block in transform domain). Since all coefficients in other quadrants of 64x64 block are being forced to zeros, the sse_diff is then a measurement of distortion introduced by such zeroing operation. 

It may not be immediately clear, for transform sizes other than 64x64, sse_diff has not been used.

Patch-set: 4