build: cmake: remove explicit CMAKE_INSTALL_PREFIX references
CMake will automatically prefix CMAKE_INSTALL_PREFIX if the
corresponding CMAKE_INSTALL_*DIR is a relative path. In the case of an
absolute path, it's generally accepted that the absolute path should be
obeyed and no prefix should be added.

In the case of openSUSE's (and Fedora's) build macros, cmake is called
with already-prefixed absolute paths for all of the CMAKE_INSTALL_*DIR
variables. These scripts work with most CMake projects, so libaom should
follow suit.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>
Change-Id: I0c7ecbcd017b35a52e2aa932b4f6a6186ce25480