Correct minimum tile width check
This check is for the minimum tile restriction in the spec, which
requires that all tiles but the rightmost should be at least 64 pixels
wide if superres not used, or 128 pixels wide if it is.

The previous check (change-id 81963) didn't do the right thing for
frames with multiple rows and just one column of tiles with a narrow
frame. This version of the code stores "min_inner_tile_width" instead
of "min_tile_width", storing -1 if there is just one column of tiles.
It also fixes up the is_min_tile_width_satisfied check with a special
case if there is just a single column.

Change-Id: I2629bbc9fa4fcf719c745124c37c7958496ba272