Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hi Rupert, I saw Deb has a fix:
> https://aomedia-review.googlesource.com/c/aom/+/83983
> Is this similar?

Hi Yunqing,

I hadn't seen that. Yes, this is essentially the same. This patch changes some names to make it a bit clearer what's going on, which I think is a little better. (With the code as it stands, a narrow frame might appear to have min_tile_width = 65536, which is a tad confusing in a debugger).

Rupert

Patch-set: 2
Reviewer: Gerrit User 7560 <7560@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>