[DO NOT SUBMIT] Show inverse ADST output range is [-32768, +32768].
This CL is NOT a proposed change. It is for experimental use only.

The range checks added in this CL for the outputs (i.e., the final
stage) of inverse ADST are not in the spec, although they are in the
spirit of the existing range checks in the spec for the earlier stages
of inverse ADST. The purpose of the new range checks is to show a
conformant 8-bit bitstream may produce an inverse ADST output value of
+32768, which cannot be represented as an int16_t.

Change-Id: I6f03f6aea3f5575541ffbdb9c4c7db7c4081dc7f