Update patch set 15
Patch Set 15:

Frank is adding some minimal coverage for this in a test in libvpx [1], it expose other overflows in the encoder so for now he's going to limit it to cases that only trigger this particular issue with bugs for followups. Once he has that in place you can pick it up in this change to help verify the fix.

[1] https://chromium-review.googlesource.com/c/webm/libvpx/+/1539258

Patch-set: 15
Reviewer: Gerrit User 5020 <5020@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>