Update patch set 5
Patch Set 5:

Hi Yunqing,
Please do not merge the current patch (as it is intermediate version). As a part of patchset 5, we reverted couple of optimizations. We are planning to root cause the issue first(among these two optimizations). After above, we shall submit the final patch with all other optimizations enabled. 
Regards,
Ranjit

Patch-set: 5