fix bad mapping of modes 26 and 27
The data structures in rdopt.c and rd.h are out of sync -- mode 26 is
{ GLOBALMV, { GOLDEN_FRAME, NONE_FRAME } }, while mode 27 is
{ GLOBALMV, { ALTREF_FRAME, NONE_FRAME } } -- however, rd.h has them
reversed. This incorrect mapping is used by the 2-pass TX search and
results in a bad mode index being copied into the best search state
seen.

This improves bd-rate by about 0.03% on the lowres test set

Change-Id: I4f6dcae261b317f3828428b2e8fb5e6655a9023a