Adjust relative distance of refs on encoder side.
get_relative_dist() sometimes mistakenly thinks that a frame is a future
reference frame even though it actually is a past reference. This is
because of limited range of order hint.

On encoder side only, we correct this value when we know it to be
impossible, based on values of lag-in-frames and arnr-max-frames.

Small improvement on a few videos:

lowres: -0.002
midres: -0.016
- station2_480p25.y4m: -0.247
- snow_mnt_480p.y4m: -0.233

STATS_CHANGED

Change-Id: Ic77ea8c6b52cd43e6f486035e1d7b007c42fe8b8