Change smooth intra disabling speed feature
This speed feature may be responsible for random seg faults
since the problems started with addition of this feature a
few weeks ago in https://aomedia-review.googlesource.com/c/aom/+/79961.

This patch changes the feature to disable smooth intra for all inter
frames. This seems to fix the issue. However the root cause may be
still related to simple motion search.

BUG=aomedia:2362

Change-Id: Ica75bcbaa41e41e4ccb467c4aea0e745ed6b56fe