Fix a segfault in simple_motion_search due to invalid search site
simple_motion_search uses NSTEP as its search_method, which depends on
search_site_config in AV1_COMP. On speed >= 6, search_site_config is not
updated correctly when there is a scaled reference frame as the
offset in search_site is not updated based on the current frame.
(search_site_config is only initialized once when AV1_COMP is created,
but the offset needs to be updated based on the current scale factor).

This commit fixes the segfault by forcing simple_motion_search to use
the search_method defined in speed feature. On speed == 6, this is set
to HEX, which does not rely on search_site_config anymore.

STATS_CHANGED

BUG=aomedia:2362

Change-Id: I10e1a165f41498413f332a46fca6cf7304f9c755